test: 拆分 Respond Agent 集成测试
